@@ -3,8 +3,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.24.2
+// 	protoc-gen-go v1.36.5
+// 	protoc        v3.19.4
 // source: base_error.proto
 
 package pb
@@ -14,6 +14,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -27,21 +28,18 @@ const (
 // 当客户端发送Request类型消息时，如果有异常则返回ErrorResponse
 // 踢下线协议也会用到该结构
 type ErrorResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Code          int32                  `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`      // 错误码
+	Message       string                 `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"` // 错误内容
 	unknownFields protoimpl.UnknownFields
-
-	Code    int32  `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`      // 错误码
-	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"` // 错误内容
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ErrorResponse) Reset() {
 	*x = ErrorResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_error_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_error_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ErrorResponse) String() string {
@@ -52,7 +50,7 @@ func (*ErrorResponse) ProtoMessage() {}
 
 func (x *ErrorResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_base_error_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -83,7 +81,7 @@ func (x *ErrorResponse) GetMessage() string {
 
 var File_base_error_proto protoreflect.FileDescriptor
 
-var file_base_error_proto_rawDesc = []byte{
+var file_base_error_proto_rawDesc = string([]byte{
 	0x0a, 0x10, 0x62, 0x61, 0x73, 0x65, 0x5f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x2e, 0x70, 0x72, 0x6f,
 	0x74, 0x6f, 0x12, 0x02, 0x70, 0x62, 0x22, 0x3d, 0x0a, 0x0d, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52,
 	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18,
@@ -91,22 +89,22 @@ var file_base_error_proto_rawDesc = []byte{
 	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65,
 	0x73, 0x73, 0x61, 0x67, 0x65, 0x42, 0x05, 0x5a, 0x03, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72,
 	0x6f, 0x74, 0x6f, 0x33,
-}
+})
 
 var (
 	file_base_error_proto_rawDescOnce sync.Once
-	file_base_error_proto_rawDescData = file_base_error_proto_rawDesc
+	file_base_error_proto_rawDescData []byte
 )
 
 func file_base_error_proto_rawDescGZIP() []byte {
 	file_base_error_proto_rawDescOnce.Do(func() {
-		file_base_error_proto_rawDescData = protoimpl.X.CompressGZIP(file_base_error_proto_rawDescData)
+		file_base_error_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_base_error_proto_rawDesc), len(file_base_error_proto_rawDesc)))
 	})
 	return file_base_error_proto_rawDescData
 }
 
 var file_base_error_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_base_error_proto_goTypes = []interface{}{
+var file_base_error_proto_goTypes = []any{
 	(*ErrorResponse)(nil), // 0: pb.ErrorResponse
 }
 var file_base_error_proto_depIdxs = []int32{
@@ -122,25 +120,11 @@ func file_base_error_proto_init() {
 	if File_base_error_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_base_error_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ErrorResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_base_error_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_base_error_proto_rawDesc), len(file_base_error_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   1,
 			NumExtensions: 0,
@@ -151,7 +135,6 @@ func file_base_error_proto_init() {
 		MessageInfos:      file_base_error_proto_msgTypes,
 	}.Build()
 	File_base_error_proto = out.File
-	file_base_error_proto_rawDesc = nil
 	file_base_error_proto_goTypes = nil
 	file_base_error_proto_depIdxs = nil
 }
