@@ -3,8 +3,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.24.2
+// 	protoc-gen-go v1.36.5
+// 	protoc        v3.19.4
 // source: player.proto
 
 package pb
@@ -14,6 +14,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -69,25 +70,22 @@ func (PlayerAttribute) EnumDescriptor() ([]byte, []int) {
 
 // 玩家基本信息
 type Player struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	PlayerId      int64                  `protobuf:"varint,1,opt,name=playerId,proto3" json:"playerId,omitempty"`     // 玩家ID
+	PlayerName    string                 `protobuf:"bytes,2,opt,name=playerName,proto3" json:"playerName,omitempty"`  // 玩家昵称
+	Gender        int32                  `protobuf:"varint,3,opt,name=gender,proto3" json:"gender,omitempty"`         // 角色性别
+	Level         int32                  `protobuf:"varint,4,opt,name=level,proto3" json:"level,omitempty"`           // 角色等级
+	CreateTime    int64                  `protobuf:"varint,5,opt,name=createTime,proto3" json:"createTime,omitempty"` // 创建时间
+	Exp           int64                  `protobuf:"varint,6,opt,name=exp,proto3" json:"exp,omitempty"`               // 经验值
 	unknownFields protoimpl.UnknownFields
-
-	PlayerId   int64  `protobuf:"varint,1,opt,name=playerId,proto3" json:"playerId,omitempty"`     // 玩家ID
-	PlayerName string `protobuf:"bytes,2,opt,name=playerName,proto3" json:"playerName,omitempty"`  // 玩家昵称
-	Gender     int32  `protobuf:"varint,3,opt,name=gender,proto3" json:"gender,omitempty"`         // 角色性别
-	Level      int32  `protobuf:"varint,4,opt,name=level,proto3" json:"level,omitempty"`           // 角色等级
-	CreateTime int64  `protobuf:"varint,5,opt,name=createTime,proto3" json:"createTime,omitempty"` // 创建时间
-	Exp        int64  `protobuf:"varint,6,opt,name=exp,proto3" json:"exp,omitempty"`               // 经验值
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Player) Reset() {
 	*x = Player{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_player_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_player_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Player) String() string {
@@ -98,7 +96,7 @@ func (*Player) ProtoMessage() {}
 
 func (x *Player) ProtoReflect() protoreflect.Message {
 	mi := &file_player_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -157,20 +155,17 @@ func (x *Player) GetExp() int64 {
 
 // 查询角色响应
 type PlayerSelectResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	List          []*Player              `protobuf:"bytes,1,rep,name=list,proto3" json:"list,omitempty"` // 玩家列表
 	unknownFields protoimpl.UnknownFields
-
-	List []*Player `protobuf:"bytes,1,rep,name=list,proto3" json:"list,omitempty"` // 玩家列表
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PlayerSelectResponse) Reset() {
 	*x = PlayerSelectResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_player_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_player_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PlayerSelectResponse) String() string {
@@ -181,7 +176,7 @@ func (*PlayerSelectResponse) ProtoMessage() {}
 
 func (x *PlayerSelectResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_player_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -205,21 +200,18 @@ func (x *PlayerSelectResponse) GetList() []*Player {
 
 // 请求创建角色
 type PlayerCreateRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	PlayerName    string                 `protobuf:"bytes,1,opt,name=playerName,proto3" json:"playerName,omitempty"` // 玩家昵称
+	Gender        int32                  `protobuf:"varint,2,opt,name=gender,proto3" json:"gender,omitempty"`        // 性别(0女,1男)
 	unknownFields protoimpl.UnknownFields
-
-	PlayerName string `protobuf:"bytes,1,opt,name=playerName,proto3" json:"playerName,omitempty"` // 玩家昵称
-	Gender     int32  `protobuf:"varint,2,opt,name=gender,proto3" json:"gender,omitempty"`        // 性别(0女,1男)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PlayerCreateRequest) Reset() {
 	*x = PlayerCreateRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_player_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_player_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PlayerCreateRequest) String() string {
@@ -230,7 +222,7 @@ func (*PlayerCreateRequest) ProtoMessage() {}
 
 func (x *PlayerCreateRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_player_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -261,20 +253,17 @@ func (x *PlayerCreateRequest) GetGender() int32 {
 
 // 创建角色回应
 type PlayerCreateResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Player        *Player                `protobuf:"bytes,1,opt,name=player,proto3" json:"player,omitempty"` // 玩家信息
 	unknownFields protoimpl.UnknownFields
-
-	Player *Player `protobuf:"bytes,1,opt,name=player,proto3" json:"player,omitempty"` // 玩家信息
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PlayerCreateResponse) Reset() {
 	*x = PlayerCreateResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_player_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_player_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PlayerCreateResponse) String() string {
@@ -285,7 +274,7 @@ func (*PlayerCreateResponse) ProtoMessage() {}
 
 func (x *PlayerCreateResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_player_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -309,20 +298,17 @@ func (x *PlayerCreateResponse) GetPlayer() *Player {
 
 // 进入游戏响应
 type PlayerEnterResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	GuideMaps     map[int32]int32        `protobuf:"bytes,2,rep,name=guideMaps,proto3" json:"guideMaps,omitempty" protobuf_key:"varint,1,opt,name=key" protobuf_val:"varint,2,opt,name=value"` // 新手引导记录列表
 	unknownFields protoimpl.UnknownFields
-
-	GuideMaps map[int32]int32 `protobuf:"bytes,2,rep,name=guideMaps,proto3" json:"guideMaps,omitempty" protobuf_key:"varint,1,opt,name=key,proto3" protobuf_val:"varint,2,opt,name=value,proto3"` // 新手引导记录列表
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PlayerEnterResponse) Reset() {
 	*x = PlayerEnterResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_player_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_player_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PlayerEnterResponse) String() string {
@@ -333,7 +319,7 @@ func (*PlayerEnterResponse) ProtoMessage() {}
 
 func (x *PlayerEnterResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_player_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -357,7 +343,7 @@ func (x *PlayerEnterResponse) GetGuideMaps() map[int32]int32 {
 
 var File_player_proto protoreflect.FileDescriptor
 
-var file_player_proto_rawDesc = []byte{
+var file_player_proto_rawDesc = string([]byte{
 	0x0a, 0x0c, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x02,
 	0x70, 0x62, 0x22, 0xa4, 0x01, 0x0a, 0x06, 0x50, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x12, 0x1a, 0x0a,
 	0x08, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x03, 0x52,
@@ -395,23 +381,23 @@ var file_player_proto_rawDesc = []byte{
 	0x65, 0x72, 0x41, 0x74, 0x74, 0x72, 0x69, 0x62, 0x75, 0x74, 0x65, 0x12, 0x14, 0x0a, 0x10, 0x41,
 	0x41, 0x5f, 0x41, 0x74, 0x74, 0x72, 0x69, 0x62, 0x75, 0x74, 0x65, 0x4e, 0x6f, 0x6e, 0x65, 0x10,
 	0x00, 0x42, 0x05, 0x5a, 0x03, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+})
 
 var (
 	file_player_proto_rawDescOnce sync.Once
-	file_player_proto_rawDescData = file_player_proto_rawDesc
+	file_player_proto_rawDescData []byte
 )
 
 func file_player_proto_rawDescGZIP() []byte {
 	file_player_proto_rawDescOnce.Do(func() {
-		file_player_proto_rawDescData = protoimpl.X.CompressGZIP(file_player_proto_rawDescData)
+		file_player_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_player_proto_rawDesc), len(file_player_proto_rawDesc)))
 	})
 	return file_player_proto_rawDescData
 }
 
 var file_player_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_player_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_player_proto_goTypes = []interface{}{
+var file_player_proto_goTypes = []any{
 	(PlayerAttribute)(0),         // 0: pb.PlayerAttribute
 	(*Player)(nil),               // 1: pb.Player
 	(*PlayerSelectResponse)(nil), // 2: pb.PlayerSelectResponse
@@ -436,73 +422,11 @@ func file_player_proto_init() {
 	if File_player_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_player_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Player); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_player_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PlayerSelectResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_player_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PlayerCreateRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_player_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PlayerCreateResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_player_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PlayerEnterResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_player_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_player_proto_rawDesc), len(file_player_proto_rawDesc)),
 			NumEnums:      1,
 			NumMessages:   6,
 			NumExtensions: 0,
@@ -514,7 +438,6 @@ func file_player_proto_init() {
 		MessageInfos:      file_player_proto_msgTypes,
 	}.Build()
 	File_player_proto = out.File
-	file_player_proto_rawDesc = nil
 	file_player_proto_goTypes = nil
 	file_player_proto_depIdxs = nil
 }
