@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.24.2
+// 	protoc-gen-go v1.36.5
+// 	protoc        v3.19.4
 // source: rpc.proto
 
 package pb
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -72,22 +73,19 @@ func (NodeState) EnumDescriptor() ([]byte, []int) {
 
 // 注册开发帐号
 type DevRegister struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	AccountName   string                 `protobuf:"bytes,1,opt,name=accountName,proto3" json:"accountName,omitempty"` // 帐号名
+	Password      string                 `protobuf:"bytes,2,opt,name=password,proto3" json:"password,omitempty"`       // 密码
+	Ip            string                 `protobuf:"bytes,3,opt,name=ip,proto3" json:"ip,omitempty"`                   // ip地址
 	unknownFields protoimpl.UnknownFields
-
-	AccountName string `protobuf:"bytes,1,opt,name=accountName,proto3" json:"accountName,omitempty"` // 帐号名
-	Password    string `protobuf:"bytes,2,opt,name=password,proto3" json:"password,omitempty"`       // 密码
-	Ip          string `protobuf:"bytes,3,opt,name=ip,proto3" json:"ip,omitempty"`                   // ip地址
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *DevRegister) Reset() {
 	*x = DevRegister{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_rpc_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_rpc_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *DevRegister) String() string {
@@ -98,7 +96,7 @@ func (*DevRegister) ProtoMessage() {}
 
 func (x *DevRegister) ProtoReflect() protoreflect.Message {
 	mi := &file_rpc_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -136,25 +134,23 @@ func (x *DevRegister) GetIp() string {
 
 // 用户信息
 type User struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	SdkId         int32                  `protobuf:"varint,1,opt,name=sdkId,proto3" json:"sdkId,omitempty"`          // sdk id
+	Pid           int32                  `protobuf:"varint,2,opt,name=pid,proto3" json:"pid,omitempty"`              // 包id
+	OpenId        string                 `protobuf:"bytes,3,opt,name=openId,proto3" json:"openId,omitempty"`         // sdk的openid
+	ServerId      int32                  `protobuf:"varint,4,opt,name=serverId,proto3" json:"serverId,omitempty"`    // 所在游戏服id
+	Uid           int64                  `protobuf:"varint,5,opt,name=uid,proto3" json:"uid,omitempty"`              // 用户id
+	Ip            string                 `protobuf:"bytes,6,opt,name=ip,proto3" json:"ip,omitempty"`                 // 请求ip
+	DeviceName    string                 `protobuf:"bytes,7,opt,name=deviceName,proto3" json:"deviceName,omitempty"` // 设备名
 	unknownFields protoimpl.UnknownFields
-
-	SdkId    int32  `protobuf:"varint,1,opt,name=sdkId,proto3" json:"sdkId,omitempty"`       // sdk id
-	Pid      int32  `protobuf:"varint,2,opt,name=pid,proto3" json:"pid,omitempty"`           // 包id
-	OpenId   string `protobuf:"bytes,3,opt,name=openId,proto3" json:"openId,omitempty"`      // sdk的openid
-	ServerId int32  `protobuf:"varint,4,opt,name=serverId,proto3" json:"serverId,omitempty"` // 所在游戏服id
-	Uid      int64  `protobuf:"varint,5,opt,name=uid,proto3" json:"uid,omitempty"`           // 用户id
-	Ip       string `protobuf:"bytes,6,opt,name=ip,proto3" json:"ip,omitempty"`              // 请求ip
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *User) Reset() {
 	*x = User{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_rpc_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_rpc_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *User) String() string {
@@ -165,7 +161,7 @@ func (*User) ProtoMessage() {}
 
 func (x *User) ProtoReflect() protoreflect.Message {
 	mi := &file_rpc_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -222,9 +218,16 @@ func (x *User) GetIp() string {
 	return ""
 }
 
+func (x *User) GetDeviceName() string {
+	if x != nil {
+		return x.DeviceName
+	}
+	return ""
+}
+
 var File_rpc_proto protoreflect.FileDescriptor
 
-var file_rpc_proto_rawDesc = []byte{
+var file_rpc_proto_rawDesc = string([]byte{
 	0x0a, 0x09, 0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0f, 0x63, 0x68, 0x65,
 	0x72, 0x72, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x70, 0x62, 0x22, 0x5b, 0x0a, 0x0b,
 	0x44, 0x65, 0x76, 0x52, 0x65, 0x67, 0x69, 0x73, 0x74, 0x65, 0x72, 0x12, 0x20, 0x0a, 0x0b, 0x61,
@@ -232,7 +235,7 @@ var file_rpc_proto_rawDesc = []byte{
 	0x52, 0x0b, 0x61, 0x63, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1a, 0x0a,
 	0x08, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52,
 	0x08, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x70, 0x18,
-	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x70, 0x22, 0x84, 0x01, 0x0a, 0x04, 0x55, 0x73,
+	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x70, 0x22, 0xa4, 0x01, 0x0a, 0x04, 0x55, 0x73,
 	0x65, 0x72, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x64, 0x6b, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28,
 	0x05, 0x52, 0x05, 0x73, 0x64, 0x6b, 0x49, 0x64, 0x12, 0x10, 0x0a, 0x03, 0x70, 0x69, 0x64, 0x18,
 	0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x03, 0x70, 0x69, 0x64, 0x12, 0x16, 0x0a, 0x06, 0x6f, 0x70,
@@ -241,27 +244,29 @@ var file_rpc_proto_rawDesc = []byte{
 	0x20, 0x01, 0x28, 0x05, 0x52, 0x08, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72, 0x49, 0x64, 0x12, 0x10,
 	0x0a, 0x03, 0x75, 0x69, 0x64, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x03, 0x75, 0x69, 0x64,
 	0x12, 0x0e, 0x0a, 0x02, 0x69, 0x70, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x70,
+	0x12, 0x1e, 0x0a, 0x0a, 0x64, 0x65, 0x76, 0x69, 0x63, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x07,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x64, 0x65, 0x76, 0x69, 0x63, 0x65, 0x4e, 0x61, 0x6d, 0x65,
 	0x2a, 0x2f, 0x0a, 0x09, 0x4e, 0x6f, 0x64, 0x65, 0x53, 0x74, 0x61, 0x74, 0x65, 0x12, 0x08, 0x0a,
 	0x04, 0x4f, 0x70, 0x65, 0x6e, 0x10, 0x00, 0x12, 0x0c, 0x0a, 0x08, 0x4d, 0x61, 0x69, 0x6e, 0x74,
 	0x61, 0x69, 0x6e, 0x10, 0x01, 0x12, 0x0a, 0x0a, 0x06, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x64, 0x10,
 	0x02, 0x42, 0x05, 0x5a, 0x03, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+})
 
 var (
 	file_rpc_proto_rawDescOnce sync.Once
-	file_rpc_proto_rawDescData = file_rpc_proto_rawDesc
+	file_rpc_proto_rawDescData []byte
 )
 
 func file_rpc_proto_rawDescGZIP() []byte {
 	file_rpc_proto_rawDescOnce.Do(func() {
-		file_rpc_proto_rawDescData = protoimpl.X.CompressGZIP(file_rpc_proto_rawDescData)
+		file_rpc_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_rpc_proto_rawDesc), len(file_rpc_proto_rawDesc)))
 	})
 	return file_rpc_proto_rawDescData
 }
 
 var file_rpc_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_rpc_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_rpc_proto_goTypes = []interface{}{
+var file_rpc_proto_goTypes = []any{
 	(NodeState)(0),      // 0: cherry.proto.pb.NodeState
 	(*DevRegister)(nil), // 1: cherry.proto.pb.DevRegister
 	(*User)(nil),        // 2: cherry.proto.pb.User
@@ -279,37 +284,11 @@ func file_rpc_proto_init() {
 	if File_rpc_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_rpc_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*DevRegister); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_rpc_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*User); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_rpc_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_rpc_proto_rawDesc), len(file_rpc_proto_rawDesc)),
 			NumEnums:      1,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -321,7 +300,6 @@ func file_rpc_proto_init() {
 		MessageInfos:      file_rpc_proto_msgTypes,
 	}.Build()
 	File_rpc_proto = out.File
-	file_rpc_proto_rawDesc = nil
 	file_rpc_proto_goTypes = nil
 	file_rpc_proto_depIdxs = nil
 }
