@@ -21,11 +21,11 @@ func (c *Component) Name() string {
 // Init 组件初始化函数
 // 为了简化部署的复杂性，本示例取消了数据库连接相关的逻辑
 func (c *Component) Init() {
+	//加在数据库配置
+	InitDatabase(c.App())
 }
 
 func (c *Component) OnAfterInit() {
-	addOnload(loadDevAccount)
-
 	for _, fn := range onLoadFuncList {
 		cherryUtils.Try(fn, func(errString string) {
 			cherryLogger.Warnf(errString)
