@@ -1,72 +1,96 @@
 package db
 
 import (
-	cherryError "github.com/cherry-game/cherry/error"
-	cherryString "github.com/cherry-game/cherry/extend/string"
-	cherryTime "github.com/cherry-game/cherry/extend/time"
-	cherryLogger "github.com/cherry-game/cherry/logger"
-	"github.com/cherry-game/examples/demo_cluster/internal/code"
-	"github.com/cherry-game/examples/demo_cluster/internal/guid"
+	"errors"
+	"fmt"
+
+	clog "github.com/cherry-game/cherry/logger"
+	"github.com/cherry-game/examples/demo_cluster/nodes/center/model"
+	"gorm.io/gorm"
 )
 
-// DevAccountTable 开发模式的帐号信息表(platform.TypeDevMode)
-type DevAccountTable struct {
-	AccountId   int64  `gorm:"column:account_id;primary_key;comment:'帐号id'" json:"accountId"`
-	AccountName string `gorm:"column:account_name;comment:'帐号名'" json:"accountName"`
-	Password    string `gorm:"column:password;comment:'密码'" json:"-"`
-	CreateIP    string `gorm:"column:create_ip;comment:'创建ip'" json:"createIP"`
-	CreateTime  int64  `gorm:"column:create_time;comment:'创建时间'" json:"createTime"`
+// AccountRepository 账户数据访问层
+type AccountRepository struct {
 }
 
-func (*DevAccountTable) TableName() string {
-	return "dev_account"
+func NewAccountRepository() *AccountRepository {
+	return &AccountRepository{}
 }
 
-func DevAccountRegister(accountName, password, ip string) int32 {
-	devAccount, _ := DevAccountWithName(accountName)
-	if devAccount != nil {
-		return code.AccountNameIsExist
+// CreateAccount 创建账户 - 简单操作，直接调用
+func CreateAccount(account *model.SlotsDevice) (*model.SlotsDevice, error) {
+	var newAccount *model.SlotsDevice
+	if err := GetDB().Create(account).First(newAccount).Error; err != nil {
+		clog.Errorf("Failed to create account: %v", err)
+		return nil, err
 	}
-
-	devAccountTable := &DevAccountTable{
-		AccountId:   guid.Next(),
-		AccountName: accountName,
-		Password:    password,
-		CreateIP:    ip,
-		CreateTime:  cherryTime.Now().Unix(),
-	}
-
-	devAccountCache.Put(accountName, devAccountTable)
-	// TODO 保存db
-
-	return code.OK
+	return newAccount, nil
 }
 
-func DevAccountWithName(accountName string) (*DevAccountTable, error) {
-	val, found := devAccountCache.GetIfPresent(accountName)
-	if found == false {
-		return nil, cherryError.Error("account not found")
+// CreateUserInfo 创建用户信息
+func CreateUserInfo(userInfo *model.SlotsUser) (*model.SlotsUser, error) {
+	var newUserInfo *model.SlotsUser
+	if err := GetDB().Create(userInfo).First(newUserInfo).Error; err != nil {
+		clog.Errorf("Failed to create user info: %v", err)
+		return nil, err
 	}
-
-	return val.(*DevAccountTable), nil
+	return newUserInfo, nil
 }
 
-// loadDevAccount 节点启动时预加载帐号数据
-func loadDevAccount() {
-	// 演示用，直接手工构建几个帐号
-	for i := 1; i <= 10; i++ {
-		index := cherryString.ToString(i)
-
-		devAccount := &DevAccountTable{
-			AccountId:   guid.Next(),
-			AccountName: "test" + index,
-			Password:    "test" + index,
-			CreateIP:    "127.0.0.1",
-			CreateTime:  cherryTime.Now().ToMillisecond(),
+// GetAccountByName 根据名称查询账户 - 简单操作，直接调用
+func GetAccountByName(accountName string) (*model.SlotsDevice, error) {
+	var account model.SlotsDevice
+	if err := GetDB().Where("device_name = ?", accountName).First(&account).Error; err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return nil, errors.New("account not found")
 		}
+		return nil, err
+	}
+	return &account, nil
+}
 
-		devAccountCache.Put(devAccount.AccountName, devAccount)
+// UpdateAccountLastLogin 更新最后登录时间 - 简单操作，直接调用
+func UpdateAccountLastLogin(accountName int64, loginTime int64, loginIP string) error {
+	return GetDB().Model(&model.SlotsDevice{}).
+		Where("device_name = ?", accountName).
+		Updates(map[string]interface{}{
+			"last_login_time": loginTime,
+			"ip_info":         loginIP,
+		}).Error
+}
+
+// GetAccountStats 获取账户统计信息 - 复杂查询，可考虑使用DB Actor
+func GetAccountStats() (*AccountStats, error) {
+	var stats AccountStats
+
+	// 总账户数
+	if err := GetDB().Model(&model.SlotsDevice{}).Count(&stats.TotalAccounts).Error; err != nil {
+		return nil, err
 	}
 
-	cherryLogger.Info("preload DevAccountTable")
+	// 今日注册数
+	if err := GetDB().Model(&model.SlotsDevice{}).
+		Where("DATE(FROM_UNIXTIME(create_time)) = CURDATE()").
+		Count(&stats.TodayRegistered).Error; err != nil {
+		return nil, err
+	}
+
+	return &stats, nil
+}
+
+// BatchCreateAccounts 批量创建账户 - 复杂操作，建议使用DB Actor
+func BatchCreateAccounts(accounts []*model.SlotsDevice) error {
+	return GetDB().Transaction(func(tx *gorm.DB) error {
+		for _, account := range accounts {
+			if err := tx.Create(account).Error; err != nil {
+				return fmt.Errorf("failed to create account %s: %v", account.DeviceName, err)
+			}
+		}
+		return nil
+	})
+}
+
+type AccountStats struct {
+	TotalAccounts   int64 `json:"totalAccounts"`
+	TodayRegistered int64 `json:"todayRegistered"`
 }
