@@ -4,6 +4,7 @@ import (
 	cslice "github.com/cherry-game/cherry/extend/slice"
 	cstring "github.com/cherry-game/cherry/extend/string"
 	cfacade "github.com/cherry-game/cherry/facade"
+	clog "github.com/cherry-game/cherry/logger"
 	"github.com/cherry-game/cherry/net/parser/pomelo"
 	pmessage "github.com/cherry-game/cherry/net/parser/pomelo/message"
 	cproto "github.com/cherry-game/cherry/net/proto"
@@ -42,7 +43,7 @@ func onPomeloDataRoute(agent *pomelo.Agent, route *pmessage.Route, msg *pmessage
 		agent.Kick(notLoginRsp, true)
 		return
 	}
-
+	//检测是不是相同的节点（相同的服务）
 	if agent.NodeType() == route.NodeType() {
 		targetPath := cfacade.NewChildPath(agent.NodeID(), route.HandleName(), session.Sid)
 		pomelo.LocalDataRoute(agent, session, route, msg, targetPath)
@@ -57,7 +58,23 @@ func gameNodeRoute(agent *pomelo.Agent, session *cproto.Session, route *pmessage
 		return
 	}
 
-	// 如果agent没有完成"角色登录",则禁止转发到game节点
+	// 1. 从session中获取玩家绑定的游戏服务器ID
+	serverId := session.GetString(sessionKey.ServerID)
+	if serverId == "" {
+		// 没有可用的游戏服务器，踢掉玩家
+		agent.Kick(&pb.Int32{Value: code.NoAvailableGameServer}, true)
+		clog.Info("player is not bind server")
+		return
+	}
+
+	// 2. 检查目标Game节点是否在线
+	if !isGameNodeOnline(agent, serverId) {
+		clog.Warnf("Player %d's bound server %s is offline, reassigning", session.Uid, serverId)
+		handleGameNodeOffline(agent, session)
+		return
+	}
+
+	// 3. 如果agent没有完成"角色登录",则禁止转发到game节点
 	if !session.Contains(sessionKey.PlayerID) {
 		// 如果不是角色登录协议则踢掉agent
 		if found := cslice.StringInSlice(msg.Route, beforeLoginRoutes); !found {
@@ -66,12 +83,55 @@ func gameNodeRoute(agent *pomelo.Agent, session *cproto.Session, route *pmessage
 		}
 	}
 
-	serverId := session.GetString(sessionKey.ServerID)
-	if serverId == "" {
-		return
-	}
-
+	// 4. 转发消息到目标游戏节点
 	childId := cstring.ToString(session.Uid)
 	targetPath := cfacade.NewChildPath(serverId, route.HandleName(), childId)
 	pomelo.ClusterLocalDataRoute(agent, session, route, msg, serverId, targetPath)
+}
+
+// 检测游戏节点是否在线
+func isGameNodeOnline(agent *pomelo.Agent, nodeID string) bool {
+	numberInfo, found := agent.Discovery().GetMember(nodeID)
+	clog.Info("game node", numberInfo)
+	return found
+}
+
+// 节点没有在线处理
+func handleGameNodeOffline(agent *pomelo.Agent, session *cproto.Session) {
+	// 1. 选择新的游戏节点
+	newGameNode := selectGameNode(agent)
+	if newGameNode == "" {
+		// 没有可用的Game节点，踢掉玩家
+		agent.Kick(&pb.Int32{Value: code.ServerMaintenance}, true)
+		return
+	}
+
+	// 2. 更新session中的serverID
+	session.Set(sessionKey.ServerID, newGameNode)
+	clog.Infof("Player %d reassigned from offline server to: %s", session.Uid, newGameNode)
+
+	// 3. 通知玩家服务器切换（可选）
+	// agent.Response(session, &pb.ReconnectResponse{
+	// 	NewServerID: newGameNode,
+	// 	Reason:      "服务器维护，已自动切换",
+	// })
+
+	// 4. 可选：保存玩家状态到数据库
+	// savePlayerStateToDatabase(session.Uid)
+}
+
+func selectGameNode(agent *pomelo.Agent) string {
+	members := agent.Discovery().ListByType("game", "")
+	if len(members) == 0 {
+		return ""
+	}
+
+	// 选择一个在线的游戏节点
+	for _, member := range members {
+		if isGameNodeOnline(agent, member.GetNodeID()) {
+			return member.GetNodeID()
+		}
+	}
+
+	return ""
 }
