@@ -1,3 +1,11 @@
+/*
+ * @Author: t 921865806@qq.com
+ * @Date: 2025-11-25 15:57:48
+ * @LastEditors: t 921865806@qq.com
+ * @LastEditTime: 2025-11-25 16:28:41
+ * @FilePath: /examples/demo_cluster/nodes/web/sdk/dev_sdk.go
+ * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+ */
 package sdk
 
 import (
@@ -35,7 +43,8 @@ func (p devSdk) Login(_ *data.SdkRow, params Params, callback Callback) {
 	}
 
 	callback(code.OK, map[string]string{
-		"open_id": cherryString.ToString(accountId),
+		"open_id":     cherryString.ToString(accountId),
+		"device_name": accountName,
 	})
 }
 
