@@ -11,7 +11,7 @@ func main() {
 		EnableWriteFile: false,
 		EnableConsole:   true,
 		MaxAge:          0,
-		TimeFormat:      "",
+		TimeFormat:      "2006-01-02 15:04:05.000",
 		PrintCaller:     false,
 	}
 
